requirements.txt
essentia-tensorflow
numpy<2